Fixed the compilation. Now compiles and runs.
Differentiated the durations of new and old types of insulin (old types renamed according to medical definitions).
Added a page for Settings, with the UI for changing the duraqtion of the . The UI isn't currently connected to ANY form o calculations..
Addad a PhysicalActivityPage, that currently is NOT in the MAUI project.
Added MAUI icon in MAUI resources, but currently not used.
Added InsulinDrug class, although currently not used.
Changed enums: TypeOfGlucoseMeasurementDevice, TypeOfGlucoseMeasurement. TypeOfInsulinSpeed