Added the work from Daniele Pieri about clicking on image position
It works, but needs to be integrated with program's data and refined in the user's interactions.
Modification of existing reference point is not present in code (TODO).
BackPicturePage and FrontPicturePage: added in the UI the buttons to save points entered by the user.
BackPicturePage: added code that manages button click in Edit and non Edit state.
FrontPicturePage: added some prototipical code to do the same (not working!)
CirclesDrawable: added code to added code to save and read the points in-from flat file.