Finished integration of Injections' positions in program and database. !! TO DEBUG and TEST !!
Database: ATTENTION: changes made
CirclesDrawable.: troubleshooting and refactories in points' visualization, added LoadCoordinates() method, to convert to internal representation the information of InsulinInjection objects
Picture pages: integrated management of graphics management of points of injections. Refactorings.
enums in Common: added enum ZoneOfPosition that represents which of the 4 images will be used in Injections' recordings.
Datalayer: added code to manage the zone in which a point is set.