Changing colors of injection points working
CirclesDrawable: added a new class CircleData to pass parameter to the Draw "automatic" function. Implemented the changing colors. Shown the selected injection with a different color.
ClickableImagePage: defined a different time for visualization of points, based on the image on which the points are drawn.
ColorConverter: added HSV color space, that was more keen to use in the injection points images.