Icon images in almost all the buttons of the previous program. Computer Sync non compiling.
FoodPage: partially added management of Manufacturer and CategoryOfFood. UI adjustments. Fix the visualization of CHO.
InjectionsPage: added four checkboxes to filter the list by image in which is stored the position. Fixed the saving of data bases on the image where the injection position are recorded.
DataLayer: fixed the saving and reading of TypeOfInjection in table Injections.
Binding:
  Increased the number of fields bound to the database (MealPage).
UI Modifications:
  TextColor = Black in all Entry controls whose color is defined manually in XAML.
  Windows' program's Window's dimensions with aspect ratio of a smart phone (modifications in MauiProgram.cs).
New Icons:
  balance, choose, chronograph_set.png, Chronograph_Started, clear, clear_fields, clock, details, escape, floppy_disc, food, get_last_blood, get_last_sugar, magnifier, minus, plus, rounding, total_sugar.