Fixed the MealPage. Diffused improvements. Introduced test project.
Injections' positions: now shows the images and the new points, but works completely only in Windows
Cosmetic modifications around the solution, several in ToolTips, that now are everywhere.
MealPage: re-done the binding and acces to database with respect with the UI.
  Now the FoodInMeal information is showm unbound in the UI and the code manages it manually.
  Information about Meal and the grid of FoodsInMeal are bound.
  The page behaves in quite acceptable mode.
RecipesPage:
  Refactored to model the MealPage. To complete and test.
UiAccuracy: fixed its behaviour when one of the two controls changes its content
Added a Unit Test project, but yet with a few methods to test.