BROKEN by AI at last moment. v.0.9.6 ready, but now not functioning.
Management of Tracking and Weighing of Food.
TrackPage: made the page that accesses the map from Open Street map and the track.
  Android: rights to access the GPS position is given, but still the phone doesn't negotiate the GPS tracking
    added the services to manage the track.
  added ancillary classes and services to the program
  BROKEN BY AI JUST BEFORE COMMITTING!!
PhysicalActivityPage:
WeighFoodPage: added Seasoning data and UI.
  deeply refactored to include b.logic in Business Layer
  added binding for data visualization
  added ancillary classes
CropPhotoPage: improved the resize and moving of the picture for cropping
Class Food: added properties to manage cooked or raw weighs (IsRaw and RawCookedRatio)
  added CRUD attributes to method to persist the new properties
Number of versions: adopted a deterministic approach for numbers of version
Business Layer and Business Object: put under the GlucoMan namespace, for uniformity with the other Layer.
Images: added a better image of the pan (pan1.svg)
AI added central package NuGet management, but maybe broke the TrackPage in the process!